+++ modules/benchmarks/src/main/scala/lucuma/catalog/votable/SimbadBenchmark.scala
@@ -1,7 +1,7 @@
 // Copyright (c) 2016-2022 Association of Universities for Research in Astronomy, Inc. (AURA)
 // For license information see LICENSE or https://opensource.org/licenses/BSD-3-Clause
 
-package lucuma.catalog
+package lucuma.catalog.votable
 
 import cats.effect._
 import cats.effect.unsafe.implicits.global
