+++ modules/catalog/src/main/scala/lucuma/catalog/votable/CatalogAdapter.scala
@@ -1,14 +1,14 @@
 // Copyright (c) 2016-2022 Association of Universities for Research in Astronomy, Inc. (AURA)
 // For license information see LICENSE or https://opensource.org/licenses/BSD-3-Clause
 
-package lucuma.catalog
+package lucuma.catalog.votable
 
 import cats.data._
 import cats.syntax.all._
 import coulomb.policy.spire.standard.given
 import coulomb.syntax.*
-import lucuma.catalog.CatalogProblem._
-import lucuma.catalog._
+import lucuma.catalog.votable.CatalogProblem._
+import lucuma.catalog.votable.*
 import lucuma.core.enums.Band
 import lucuma.core.enums.CatalogName
 import lucuma.core.math.BrightnessUnits._
