+++ modules/catalog/src/main/scala/lucuma/catalog/votable/package.scala
@@ -1,14 +1,14 @@
 // Copyright (c) 2016-2022 Association of Universities for Research in Astronomy, Inc. (AURA)
 // For license information see LICENSE or https://opensource.org/licenses/BSD-3-Clause
 
-package lucuma.catalog
+package lucuma.catalog.votable
 
 import cats.data._
 import cats.syntax.all._
 import eu.timepit.refined._
 import eu.timepit.refined.collection.NonEmpty
 import eu.timepit.refined.types.string.NonEmptyString
-import lucuma.catalog.CatalogProblem.FieldValueProblem
+import lucuma.catalog.votable.CatalogProblem.FieldValueProblem
 import lucuma.core.syntax.string._
 
 def parseDoubleValue(
