+++ modules/testkit/src/main/scala/lucuma/catalog/votable/arb/ArbUcd.scala
@@ -1,12 +1,12 @@
 // Copyright (c) 2016-2022 Association of Universities for Research in Astronomy, Inc. (AURA)
 // For license information see LICENSE or https://opensource.org/licenses/BSD-3-Clause
 
-package lucuma.catalog.arb
+package lucuma.catalog.votable.arb
 
 import cats.data.NonEmptyList
 import eu.timepit.refined.scalacheck.string._
 import eu.timepit.refined.types.string.NonEmptyString
-import lucuma.catalog._
+import lucuma.catalog.votable.*
 import org.scalacheck.Cogen
 import org.scalacheck.Gen
 import org.scalacheck._
